# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with PHP Laravel",
        "description": "Integrate Swagger in Laravel application",
        "contact": {
            "email": "afrozsatvilkar2014@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/forgotpassword": {
            "post": {
                "summary": "forgot password",
                "description": "forgot user password",
                "operationId": "99e2b5ebccc1ce3536f8c535b06b8911",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "we have mailed your password reset link to respective E-mail"
                    },
                    "404": {
                        "description": "we can not find a user with that email address"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/resetpassword": {
            "post": {
                "summary": "reset password",
                "description": "reset user password",
                "operationId": "7b46729d37c5686d0367d86cb409c57e",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "new_password",
                                    "confirm_password"
                                ],
                                "properties": {
                                    "new_password": {
                                        "type": "password"
                                    },
                                    "confirm_password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfull!"
                    },
                    "400": {
                        "description": "we can't find the user with that e-mail address"
                    },
                    "401": {
                        "description": "This token is invalid"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}